Widen relay seq to 12 bits and move it into the payload (#10)
Random 8-bit nonces collided too often in the dedup ring. Widen the sequence
number to 12 bits using the 4 spare payload bits, and move it out of the
manufacturer-data header into the payload (bits 116-127) so the mesh and macOS
paths carry and read it identically.

- Header drops its seq byte (6->5 bytes); seq now lives in the payload via
  message_get_seq/message_set_seq (12-bit).
- Mesh originate goes back to an incrementing counter (g_seq, wraps at 0xFFF)
  so back-to-back triggers never collide. Dedup is unchanged — still a pure
  existence check, making no assumption the counter increases, so a reset that
  restarts it at 0 is harmless beyond a brief stale-entry collision chance.
- The two RX paths share a new dedup_and_act() helper; the macOS gateway packet
  no longer writes a header seq byte.
- broadcast_log seq widened uint8_t -> uint16_t through core/wrapper/headers.
- bleatctl.py mirrors the 12-bit payload seq (random nonce — the Mac is
  stateless per run and dedup only checks existence).
- tests: 12-bit seq round-trip, no bleed into slots/command, wide-seq dedup.

Breaking wire change — the whole fleet must run this firmware together.

+++ app/src/broadcast_log.h
@@ -15,12 +15,12 @@ void broadcast_log_init(void);
  *
  * @return true if the pair was already in the log (duplicate).
  */
-bool broadcast_log_check_and_record(uint8_t originator, uint8_t seq);
+bool broadcast_log_check_and_record(uint8_t originator, uint16_t seq);
 
 /**
  * Unconditionally record (originator, seq) in the log.
  * Use when the caller knows the entry is new (e.g. its own broadcast).
  */
-void broadcast_log_record(uint8_t originator, uint8_t seq);
+void broadcast_log_record(uint8_t originator, uint16_t seq);
 
 #endif /* BROADCAST_LOG_H_ */
